requirements.txt
google-adk>=1.15.0
google-cloud-aiplatform[evaluation,agent-engines]
protobuf
arxiv
gtts
python-pptx
moviepy
Pillow
fastapi
uvicorn
python-dotenv
firebase-admin
google-cloud-firestore
httpx
pyjwt[crypto]
apscheduler
pypdf
google-cloud-storage